# lib/api-zod/src/generated/types/jobStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusResponseStatus } from "./jobStatusResponseStatus";

export interface JobStatusResponse {
  jobId: string;
  status: JobStatusResponseStatus;
  progress: string;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  startTime?: number | null;
}